edit NameAPI
